+++ serverscripts/nginx.py
@@ -1,7 +1,7 @@
 """Extract information from nginx config files.
 
 """
-from serverscripts.six.moves.urllib.parse import urlparse
+from urllib.parse import urlparse
 
 import argparse
 import copy