+++ serverscripts/tests/test_docker.py
@@ -16,6 +16,7 @@ cca6ed94102fa749cd32cb289cde07d38323380697a87f6c6334de9715caac76	harbor.lizard.n
 9d6dda1ad3054870b157252c3ceb752a67d5382430b8d37a9d40b72db83f2777	minio/minio:RELEASE.2020-10-03T02-19-42Z	"/usr/bin/docker-entrypoint.sh server /export"	2021-04-15 13:42:06 +0200 CEST	2 weeks ago	0.0.0.0:9000->9000/tcp, :::9000->9000/tcp	Up 51 minutes	0B (virtual 62.4MB)	threedi-api_minio_1	1c18c1163d44b55769f5090b8ae12fff8c3f970f63007d98b5130b1fdaaebab5,threedi-api_miniodata	threedi_backend
 """
 
+
 class DockerTestCase(TestCase):
     @mock.patch("serverscripts.docker.get_output")
     @mock.patch("serverscripts.docker.container_details")
@@ -32,5 +33,7 @@ class DockerTestCase(TestCase):
         mock_output.return_value = (DOCKER_PS_OUTPUT, "")
         result = docker.container_details()
         self.assertEqual(2, len(result))
-        self.assertEqual(result[0]["command"], "\"python manage.py runserver 0.0.0.0:8000\"")
+        self.assertEqual(
+            result[0]["command"], '"python manage.py runserver 0.0.0.0:8000"'
+        )
         self.assertEqual(result[1]["size"], "0B (virtual 62.4MB)")